Move net::udp tests to their own files
